checked we are adding offset in correct place

+++ src/drivers/vmount/output.cpp
@@ -250,7 +250,7 @@ void OutputBase::_calculate_output_angles(const hrt_abstime &t)
 		_change_in_alt = change_in_altitude;
 		_correction = change_in_altitude * altitude_factor;
 //		_angle_setpoints[1] = _cur_control_data->type_data.angle_gradient.pitch + change_in_altitude * altitude_factor;
-		_angle_setpoints[1] = _cur_control_data->type_data.angle_gradient.pitch + 15 * M_DEG_TO_RAD;
+		_angle_setpoints[1] = _cur_control_data->type_data.angle_gradient.pitch + (float)(15.0 * M_DEG_TO_RAD);
 		_angle_setpoints[2] = 0.f;
 	}
